Refactor homepage with add lead form
